+++ packages/brick_supabase/test/__mocks__.dart
@@ -1,6 +1,8 @@
 import 'package:brick_supabase/brick_supabase.dart';
 
-class DemoModel extends SupabaseModel {
+part '__mocks_generated__.dart';
+
+class Demo extends SupabaseModel {
   @Supabase(unique: true)
   final String id;
 
@@ -8,68 +10,11 @@ class DemoModel extends SupabaseModel {
 
   final int age;
 
-  DemoModel({
+  Demo({
     required this.id,
     required this.name,
     required this.age,
   });
-}
-
-DemoModel _$DemoModelFromSupabase(Map<String, dynamic> json) {
-  return DemoModel(
-    id: json['id'] as String,
-    name: json['name'] as String,
-    age: json['age'] as int,
-  );
-}
-
-Future<Map<String, dynamic>> _$DemoModelToSupabase(DemoModel instance) async {
-  return <String, dynamic>{
-    'id': instance.id,
-    'name': instance.name,
-    'age': instance.age,
-  };
-}
-
-class DemoModelAdapter extends SupabaseAdapter<DemoModel> {
-  @override
-  Future<DemoModel> fromSupabase(data, {required provider, repository}) async =>
-      _$DemoModelFromSupabase(data);
-
-  @override
-  Future<Map<String, dynamic>> toSupabase(instance, {required provider, repository}) async =>
-      await _$DemoModelToSupabase(instance);
-
-  @override
-  final defaultToNull = false;
-
-  @override
-  final fieldsToSupabaseColumns = {
-    'id': const RuntimeSupabaseColumnDefinition(
-      association: false,
-      columnName: 'id',
-    ),
-    'name': const RuntimeSupabaseColumnDefinition(
-      association: false,
-      columnName: 'name',
-    ),
-    'age': const RuntimeSupabaseColumnDefinition(
-      association: false,
-      columnName: 'age',
-    ),
-  };
-
-  @override
-  final ignoreDuplicates = true;
-
-  @override
-  final onConflict = null;
-
-  @override
-  final supabaseTableName = 'demos';
-
-  @override
-  final uniqueFields = {'id'};
 }
 
 class DemoNestedAssociationModel extends SupabaseModel {
@@ -78,7 +23,7 @@ class DemoNestedAssociationModel extends SupabaseModel {
 
   final String name;
 
-  @Supabase(name: 'nested_id')
+  @Supabase(name: 'nested_column')
   final DemoAssociationModel nested;
 
   DemoNestedAssociationModel({
@@ -88,72 +33,12 @@ class DemoNestedAssociationModel extends SupabaseModel {
   });
 }
 
-DemoNestedAssociationModel _$DemoNestedAssociationModelFromSupabase(Map<String, dynamic> json) {
-  return DemoNestedAssociationModel(
-    id: json['id'] as String,
-    name: json['name'] as String,
-    nested: _$DemoAssociationModelFromSupabase(json['nested'] as Map<String, dynamic>),
-  );
-}
-
-Future<Map<String, dynamic>> _$DemoNestedAssociationModelToSupabase(
-  DemoNestedAssociationModel instance,
-) async {
-  return <String, dynamic>{
-    'id': instance.id,
-    'name': instance.name,
-    'nested': await _$DemoAssociationModelToSupabase(instance.nested),
-  };
-}
-
-class DemoNestedAssociationModelAdapter extends SupabaseAdapter<DemoNestedAssociationModel> {
-  @override
-  Future<DemoNestedAssociationModel> fromSupabase(data, {required provider, repository}) async =>
-      _$DemoNestedAssociationModelFromSupabase(data);
-
-  @override
-  Future<Map<String, dynamic>> toSupabase(instance, {required provider, repository}) async =>
-      await _$DemoNestedAssociationModelToSupabase(instance);
-
-  @override
-  final defaultToNull = false;
-
-  @override
-  final fieldsToSupabaseColumns = {
-    'id': const RuntimeSupabaseColumnDefinition(
-      association: false,
-      columnName: 'id',
-    ),
-    'name': const RuntimeSupabaseColumnDefinition(
-      association: false,
-      columnName: 'name',
-    ),
-    'nested': const RuntimeSupabaseColumnDefinition(
-      association: true,
-      columnName: 'nested_id',
-      associationType: DemoAssociationModel,
-    ),
-  };
-
-  @override
-  final ignoreDuplicates = true;
-
-  @override
-  final onConflict = null;
-
-  @override
-  final supabaseTableName = 'demo_nested_associations';
-
-  @override
-  final uniqueFields = {'id'};
-}
-
 class DemoAssociationModel extends SupabaseModel {
   @Supabase(name: 'assocs_id')
-  final List<DemoModel>? assocs;
+  final List<Demo>? assocs;
 
   @Supabase(name: 'assoc_id')
-  final DemoModel assoc;
+  final Demo assoc;
 
   @Supabase(unique: true)
   final String id;
@@ -168,87 +53,36 @@ class DemoAssociationModel extends SupabaseModel {
   });
 }
 
-DemoAssociationModel _$DemoAssociationModelFromSupabase(Map<String, dynamic> json) {
-  return DemoAssociationModel(
-    id: json['id'] as String,
-    name: json['name'] as String,
-    assoc: _$DemoModelFromSupabase(json['assoc'] as Map<String, dynamic>),
-  );
+class RecursiveParent extends SupabaseModel {
+  final RecursiveChild child;
+
+  final String parentId;
+
+  RecursiveParent({
+    required this.child,
+    required this.parentId,
+  });
 }
 
-Future<Map<String, dynamic>> _$DemoAssociationModelToSupabase(
-  DemoAssociationModel instance, {
-  provider,
-  repository,
-}) async {
-  return <String, dynamic>{
-    'id': instance.id,
-    'name': instance.name,
-    'assocs': await Future.wait<Map<String, dynamic>>(
-      instance.assocs
-              ?.map(
-                (s) => DemoModelAdapter().toSupabase(s, provider: provider, repository: repository),
-              )
-              .toList() ??
-          [],
-    ),
-    'assoc': await _$DemoModelToSupabase(instance.assoc),
-  };
-}
+class RecursiveChild extends SupabaseModel {
+  final RecursiveParent parent;
 
-class DemoAssociationModelAdapter extends SupabaseAdapter<DemoAssociationModel> {
-  @override
-  Future<DemoAssociationModel> fromSupabase(data, {required provider, repository}) async =>
-      _$DemoAssociationModelFromSupabase(data);
+  final String childId;
 
-  @override
-  Future<Map<String, dynamic>> toSupabase(instance, {required provider, repository}) async =>
-      await _$DemoAssociationModelToSupabase(instance);
+  final Demo otherAssoc;
 
-  @override
-  final defaultToNull = false;
-
-  @override
-  final fieldsToSupabaseColumns = {
-    'id': const RuntimeSupabaseColumnDefinition(
-      association: false,
-      columnName: 'id',
-    ),
-    'name': const RuntimeSupabaseColumnDefinition(
-      association: false,
-      columnName: 'name',
-    ),
-    'assoc': const RuntimeSupabaseColumnDefinition(
-      association: true,
-      associationIsNullable: false,
-      columnName: 'assoc_id',
-      foreignKey: 'assoc_id',
-      associationType: DemoModel,
-    ),
-    'assocs': const RuntimeSupabaseColumnDefinition(
-      association: true,
-      associationIsNullable: true,
-      columnName: 'assocs_id',
-      associationType: DemoModel,
-    ),
-  };
-
-  @override
-  final ignoreDuplicates = true;
-
-  @override
-  final onConflict = null;
-
-  @override
-  final supabaseTableName = 'demo_associations';
-
-  @override
-  final uniqueFields = {'id'};
+  RecursiveChild({
+    required this.parent,
+    required this.childId,
+    required this.otherAssoc,
+  });
 }
 
 final Map<Type, SupabaseAdapter<SupabaseModel>> _supabaseMappings = {
-  DemoModel: DemoModelAdapter(),
+  Demo: DemoAdapter(),
   DemoAssociationModel: DemoAssociationModelAdapter(),
   DemoNestedAssociationModel: DemoNestedAssociationModelAdapter(),
+  RecursiveParent: RecursiveParentAdapter(),
+  RecursiveChild: RecursiveChildAdapter(),
 };
 final supabaseModelDictionary = SupabaseModelDictionary(_supabaseMappings);